Use build tags instead of runtime.GOARCH to select ISA
